make removeRelateds and pruneVars optional
May want to run this workflow on files where related samples and
variants in LD have already been removed, so make these steps
optional. This requires outputting bed/bim/fam files from those
functions instead of lists of samples and variants, so a bed file
can be passed to the next step whether or not pruning has happened,
etc. Keeping track of whether of not to apply sample and variant
filters in subsequent steps is much more complicated, and bed
files will be copied to a new instance for each task regardless.