use pvar to check overlap instead of subsetting loadings and freqs
If we skip the subset step for the loading and freq files, the
format is much more flexible because plink can read column headers,
but the awk commands in prepareFiles required the variant ID in a
specific column position. Instead, we can use the .pvar file to
count the number of variants, by modifying the python code to
skip header lines when counting variants.