remove any snps that are monomorphic before creating PCs
got an error from run_pca_projected on the same dataset as was
used to create the pcs:
Error: --score variance-standardize failure for variant '1:14848791:G:A':
estimated allele frequency is zero or NaN, but not all dosages are zero. (This
is possible when e.g. allele frequencies are estimated from founders, but the
allele is only observed in nonfounders.)
Since we selected SNPs and then removed relateds, need an extra step
to remove any snps that are now monomorphic